\name{as.SpatialPointsDataFrame}
\alias{as.SpatialPointsDataFrame}

\title{
Function to convert grid objects to SpatialPointsDataFrame objects
}
\description{
This function converts \code{\link{grid}} objects defined in the \pkg{grid} package to objects of class \code{\link{SpatialPointsDataFrame}} defined in \pkg{sp}.
}
\usage{
as.SpatialPointsDataFrame(x)
}
\arguments{
  \item{x}{A \code{\link{grid}} object.}
}
\value{
An object of class \code{\link{SpatialPointsDataFrame}}
}
\author{
Ethan Plunkett
}

\seealso{
\code{\link{as.grid}}
}
\examples{
#
}
\keyword{ spatial }
